third_party/blink/web_tests/external/wpt/app-history/navigate/navigate-without-url-arg.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    let navInfo = { nav: "info" };
    let start_url = i.contentWindow.location.href;
    let start_key = i.contentWindow.appHistory.current.key;
    let start_id = i.contentWindow.appHistory.current.id;
    let onnavigate_called = false;
    i.contentWindow.appHistory.onnavigate = t.step_func(e => {
      onnavigate_called = true;
      assert_equals(e.info, navInfo);
    });
    i.contentWindow.appHistory.navigate({ info: navInfo });
    i.onload = t.step_func_done(() => {
      assert_true(onnavigate_called);
      assert_equals(start_url, i.contentWindow.location.href);
      assert_equals(start_key, i.contentWindow.appHistory.current.key);
      assert_not_equals(start_id, i.contentWindow.appHistory.current.id);
      assert_equals(i.contentWindow.appHistory.entries().length, 1);
    });
  });
}, "navigate() variant with no url");
</script>
